fix: price.close 타입을 decimal에서 int로 변경

+++ script/run_pykrx.py
@@ -60,7 +60,7 @@ def load_price():
                 {
                     "ticker": ticker,
                     "date":   idx.strftime("%Y-%m-%d"),
-                    "close":  float(row["종가"]),
+                    "close":  int(row["종가"]),
                     "volume": int(row["거래량"]),
                 }
                 for idx, row in df.iterrows()
